# frontend/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Custody Scheduler API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/schedule/": {
      "get": {
        "summary": "Get Schedule",
        "operationId": "get_schedule_api_v1_schedule__get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyCustodyState"
                  },
                  "title": "Response Get Schedule Api V1 Schedule  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/feed.ics": {
      "get": {
        "summary": "Get Calendar Feed",
        "description": "Subscribeable ICS feed authenticated by opaque calendar_feed_token.",
        "operationId": "get_calendar_feed_api_v1_schedule_feed_ics_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/overrides/pending": {
      "get": {
        "summary": "List Pending Overrides",
        "operationId": "list_pending_overrides_api_v1_schedule_overrides_pending_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleOverride"
                  },
                  "title": "Response List Pending Overrides Api V1 Schedule Overrides Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/overrides": {
      "post": {
        "summary": "Create Override",
        "operationId": "create_override_api_v1_schedule_overrides_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleOverride"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOverride"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/overrides/{override_id}/decision": {
      "post": {
        "summary": "Decide Override Request",
        "operationId": "decide_override_request_api_v1_schedule_overrides__override_id__decision_post",
        "parameters": [
          {
            "name": "override_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Override Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOverride"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/overrides/sweep-expired": {
      "post": {
        "summary": "Sweep Expired Overrides",
        "operationId": "sweep_expired_overrides_api_v1_schedule_overrides_sweep_expired_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Sweep Expired Overrides Api V1 Schedule Overrides Sweep Expired Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyCustodyState": {
        "properties": {
          "current_date": {
            "type": "string",
            "format": "date",
            "title": "Current Date"
          },
          "baseline_parent": {
            "$ref": "#/components/schemas/ParentRole"
          },
          "final_parent": {
            "$ref": "#/components/schemas/ParentRole"
          },
          "is_overridden": {
            "type": "boolean",
            "title": "Is Overridden"
          },
          "override_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleOverride"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "current_date",
          "baseline_parent",
          "final_parent",
          "is_overridden"
        ],
        "title": "DailyCustodyState"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OverrideDecisionRequest": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "OverrideDecisionRequest"
      },
      "OverrideStatus": {
        "type": "string",
        "enum": [
          "Draft",
          "Pending",
          "Approved",
          "Rejected",
          "Expired"
        ],
        "title": "OverrideStatus"
      },
      "OverrideType": {
        "type": "string",
        "enum": [
          "Holiday",
          "Mutual Swap",
          "Emergency"
        ],
        "title": "OverrideType"
      },
      "ParentRole": {
        "type": "string",
        "enum": [
          "Parent A",
          "Parent B"
        ],
        "title": "ParentRole"
      },
...
      "ScheduleOverride": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "override_date": {
            "type": "string",
            "format": "date",
            "title": "Override Date"
          },
          "assigned_parent": {
            "$ref": "#/components/schemas/ParentRole"
          },
          "override_type": {
            "$ref": "#/components/schemas/OverrideType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "status": {
            "$ref": "#/components/schemas/OverrideStatus",
            "default": "Approved"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "requested_by_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By User Id"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "requested_by_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Label"
          }
        },
        "type": "object",
        "required": [
          "override_date",
          "assigned_parent",
          "override_type",
          "description"
        ],
        "title": "ScheduleOverride"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}